迁移str2Xml()函数到utils模块，并添加单元测试
